Assert that the threaded mixer isn't spinning on empty requests
Now that the mixer is threaded, we should detect when programmatic
empty requests are made and fix them to prevent runaway spinning.

The functions themselves already handle the "empty request" case,
so there's no harm or risk with release builds. But we might as
well catch these in debug builds.