Eliminate the hidden `InstantLaunch` verbosity level
The InstantLaunch verbosity level wasn't actually settable
by the user, and seemed was doubling in purpose to indicate
if the user was launching an executable directly from the
DOSBox command line.

This also caused a problem where the `quiet` verbosity
level was incorrectly used when it should be `low` per
the `startup_verbosity` conf documentation.